% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_log_id.R
\name{get_param_log_id}
\alias{get_param_log_id}
\title{Get log ID from parameter list}
\usage{
get_param_log_id(param_list, in_dir_path)
}
\arguments{
\item{param_list}{named list of parameter values. Each element should be a parameter name - value pair. Values should be atomic vectors or NULL.}

\item{in_dir_path}{character vector of length 1 representing the input directory the params.log file should be read from.}
}
\value{
the log ID matching <param_list> in the params.log file at <in_dir_path>.
}
\description{
Get log ID from parameter list
}
\examples{
# Define analysis parameters list
param_list <- list(input_count = "tpm_lengthScaledTPM",
                   model_formula = "~ n_gene_on + exposure + (1 | line_name)",
                   cell_type = "neuron",
                   lines = c("B856", "B156", "B067"),
                   protein_coding_only = TRUE,
                   min_cnt_excl = 0,
                   min_freq_incl = 0.2,
                   padj = 0.05)

# Write parameters log file
write_param_log(param_list = param_list, out_dir_path = "~")

# Get log ID from parameters list
param_log_id <- get_param_log_id(param_list = param_list, in_dir_path = "~")

# Delete example parameters log file
file.remove("~/params.log")

}
