% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_log_id_from_names.R
\name{get_param_log_id_from_names}
\alias{get_param_log_id_from_names}
\title{Get log ID from parameter names}
\usage{
get_param_log_id_from_names(params, in_dir_path)
}
\arguments{
\item{params}{character vector of object names. All objects should be defined in the global environment and bound to atomic vectors or NULL.}

\item{in_dir_path}{character vector of length 1 representing the input directory the params.log file should be read from.}
}
\value{
the log ID matching \if{html}{\out{<params>}} in the params.log file at <in_dir_path>.
}
\description{
Get log ID from parameter names
}
\examples{
# Define analysis parameters
input_count <- "tpm_lengthScaledTPM"
model_formula <- "~ n_gene_on + exposure + (1 | line_name)"
cell_type <- "neuron"
lines <- c("B856", "B156", "B067")
protein_coding_only <- TRUE
min_cnt_excl <- 0
min_freq_incl <- 0.2
padj <- 0.05

params <- c("input_count", "model_formula", "cell_type", "lines",
            "protein_coding_only", "min_cnt_excl", "min_freq_incl", "padj")

# Write parameters log file
write_param_log_from_names(params = params, out_dir_path = "~")

# Get log ID from parameters values
param_log_id <- get_param_log_id_from_names(params = params, in_dir_path = "~")

# Delete example parameters log file
file.remove("~/params.log")

}
