% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_log_id_from_list.R
\name{get_params_log_id_from_list}
\alias{get_params_log_id_from_list}
\title{Get log ID from parameters list}
\usage{
get_params_log_id_from_list(params_list, in_dir_path)
}
\arguments{
\item{params_list}{named list of parameter values. Each element should be a parameter name - value pair. Values should be atomic vectors or NULL.}

\item{in_dir_path}{character vector of length 1 representing the input directory the params.log file should be read from.}
}
\value{
the log ID matching <params_list> in the params.log file at <in_dir_path>.
}
\description{
Get log ID from parameters list
}
\examples{
# Define analysis parameters list
params_list <- list(input_count = "tpm_lengthScaledTPM",
                    model_formula = "~ n_gene_on + exposure + (1 | line_name)",
                    cell_type = "neuron",
                    lines = c("B856", "B156", "B067"),
                    protein_coding_only = TRUE,
                    min_cnt_excl = 0,
                    min_freq_incl = 0.2,
                    padj = 0.05)

# Retrieve parameters names
params_names <- names(params_list)

# Assign parameters names to values in global environment
for(i in params_names) {
  assign(x = i, value = params_list[[i]])
}

# Write parameters log file
write_params_log(params = params_names, out_dir_path = "~")

# Get log ID from parameters list
params_log_id <- get_params_log_id_from_list(params_list = params_list, in_dir_path = "~")

# Delete example parameters log file
file.remove("~/params.log")

}
