% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_params_log.R
\name{write_params_log}
\alias{write_params_log}
\title{Write parameters log file}
\usage{
write_params_log(params, out_dir_path, params_log_table = FALSE)
}
\arguments{
\item{params}{data.frame or tibble with character column <gene_id> and double columns representing sample-specific gene counts.}

\item{out_dir_path}{character vector of length 1 or name representing gene ID. Must be a column name of input.}

\item{params_log_table}{data.frame or tibble with character / factor columns <sample_id_var> and <group_id_vars>.}
}
\value{
a tibble with number of cells, number of expressing cells and proportion of expressing cells per biological group ID.
}
\description{
Write parameters log file
}
\examples{
# Define analysis parameters
input_count <- "tpm_lengthScaledTPM"
model_formula <- "~ n_gene_on + exposure + (1 | line_name)"
cell_type <- "neuron"
lines <- c("B856", "B156", "B067")
protein_coding_only <- TRUE
min_cnt_excl <- 0
min_freq_incl <- 0.2
padj <- 0.05

# Write parameters log file
write_params_log(params = c("input_count", "model_formula", "cell_type", "lines",
                            "protein_coding_only", "min_cnt_excl", "min_freq_incl", "padj"),
                 out_dir_path = "~")

}
