% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_params_log_from_list.R
\name{write_params_log_from_list}
\alias{write_params_log_from_list}
\title{Write parameters log file from parameters list}
\usage{
write_params_log_from_list(params_list, out_dir_path, params_log_table = TRUE)
}
\arguments{
\item{params_list}{named list of parameter values. Each element should be a parameter name - value pair. Values should be atomic vectors or NULL.}

\item{out_dir_path}{character vector of length 1 representing the output directory where the params.log file should be stored.}

\item{params_log_table}{logical vector of length 1 indicating if the params log table should be returned.}
}
\value{
a params.log file written in <out_dir_path> and optionally the corresponding params log table.
}
\description{
Write parameters log file from parameters list
}
\examples{
# Define params list
params_list <- list(input_count = "tpm_lengthScaledTPM",
                    model_formula = "~ n_gene_on + exposure + (1 | line_name)",
                    cell_type = "neuron",
                    lines = c("B856", "B156", "B067"),
                    protein_coding_only = TRUE,
                    min_cnt_excl = 0,
                    min_freq_incl = 0.2,
                    padj = 0.05)

# Write parameters log file
write_params_log_from_list(params_list = params_list, out_dir_path = "~")

# Delete example parameters log file
file.remove("~/params.log")

}
