% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_params_log.R
\name{write_params_log}
\alias{write_params_log}
\title{Write parameters log file}
\usage{
write_params_log(params, out_dir_path, params_log_table = TRUE)
}
\arguments{
\item{params}{character vector of object names. All objects should be defined in the global environment and bound to atomic vectors.}

\item{out_dir_path}{character vector of length 1 representing the output directory where the params.log file should be stored.}

\item{params_log_table}{logical vector of length 1 indicating if the params log table should be returned.}
}
\value{
a params.log file written in <out_dir_path> and optionally the corresponding params log table.
}
\description{
Write parameters log file
}
\examples{
# Define analysis parameters
input_count <- "tpm_lengthScaledTPM"
model_formula <- "~ n_gene_on + exposure + (1 | line_name)"
cell_type <- "neuron"
lines <- c("B856", "B156", "B067")
protein_coding_only <- TRUE
min_cnt_excl <- 0
min_freq_incl <- 0.2
padj <- 0.05

params <- c("input_count", "model_formula", "cell_type", "lines",
            "protein_coding_only", "min_cnt_excl", "min_freq_incl", "padj")

# Write parameters log file
write_params_log(params = params, out_dir_path = "~")

# Delete example parameters log file
file.remove("~/params.log")

}
